RA-1244 - :construction: Moving mapping related tests under the mappings folder in the unit test project.
