+++ src/Esfa.Vacancy.Register.UnitTests/Api/Mappings/ApprenticeshipMapper/WhenGettingLiveVacancyWithLegacyTextWageType.cs
@@ -11,18 +11,17 @@ using Moq;
 using NUnit.Framework;
 using Ploeh.AutoFixture;
 
-namespace Esfa.Vacancy.Register.UnitTests.Api.Orchestrators.VacancyOrchestratorTests
+namespace Esfa.Vacancy.Register.UnitTests.Api.Mappings.ApprenticeshipMapper
 {
     [TestFixture]
-    public class WhenGettingLiveVacancyWithAmbiguousWage
+    public class WhenGettingLiveVacancyWithLegacyTextWageType
     {
-        [TestCase(WageType.Unwaged, "Unwaged")]
-        [TestCase(WageType.ToBeAgreedUponAppointment, "To be agreed upon appointment")]
-        [TestCase(WageType.CompetitiveSalary, "Competitive salary")]
-        public async Task ShouldHaveAppropriateWageDescription(WageType wageType, string expectedWageText)
+        [Test]
+        public async Task ShouldHaveUnknownWageForVacanciesWithLegacyTextWageType()
         {
             const int vacancyReference = 1234;
             const int liveVacancyStatusId = 2;
+            const string unknownwWageText = "Unknown";
 
             var mockMediator = new Mock<IMediator>();
             var provideSettings = new Mock<IProvideSettings>();
@@ -35,8 +34,7 @@ namespace Esfa.Vacancy.Register.UnitTests.Api.Orchestrators.VacancyOrchestratorT
                                             .With(v => v.VacancyReferenceNumber, vacancyReference)
                                             .With(v => v.VacancyStatusId, liveVacancyStatusId)
                                             .With(v => v.VacancyTypeId, (int)VacancyType.Apprenticeship)
-                                            .With(v => v.WageType, (int)wageType)
-                                            .Without(v => v.WeeklyWage)
+                                            .With(v => v.WageType, (int)WageType.LegacyText)
                                             .Without(v => v.WageUnitId)
                                             .Create()
                 });
@@ -45,7 +43,7 @@ namespace Esfa.Vacancy.Register.UnitTests.Api.Orchestrators.VacancyOrchestratorT
 
             result.VacancyReference.Should().Be(vacancyReference);
             result.WageUnit.Should().BeNull();
-            result.WageText.Should().Be(expectedWageText);
+            result.WageText.Should().Be(unknownwWageText);
         }
     }
 }
